remove useToast hook and move functions to utils
