+++ components/tables/edit-cards-table/index.tsx
@@ -1,7 +1,7 @@
 import ButtonGroup from '@components/buttons/button-group'
 import SearchBar from '@components/inputs/search-bar'
 import EditCardModal from '@components/modals/edit-card-modal'
-import { warningToast } from '@hooks/use-toast'
+import { warningToast } from '@utils/toasts'
 import { useEditCard } from '@pages/api/mutations'
 import React, { useMemo, useState } from 'react'
 import {
