+++ components/tables/issue-packs-table/index.tsx
@@ -10,7 +10,7 @@ import Table from '../table'
 import React, { useMemo, useState } from 'react'
 import useIssuePack from '@pages/api/mutations/use-issue-pack'
 import IssuePacksModal from '@components/modals/issue-packs-modal'
-import { warningToast } from '@hooks/use-toast'
+import { warningToast } from '@utils/toasts'
 import getUidFromSession from '@utils/get-uid-from-session'
 import packsMap from '@constants/packs-map'
