+++ components/tables/process-cards-table/index.tsx
@@ -11,7 +11,7 @@ import Pagination from '../pagination'
 import Table from '../table'
 import ProcessCardModal from '@components/modals/process-card-modal'
 import { useAcceptCard, useDenyCard } from '@pages/api/mutations'
-import { warningToast } from '@hooks/use-toast'
+import { warningToast } from '@utils/toasts'
 import attributesMap from '@constants/attributes-map'
 
 type ProcessCardsTableProps = {
