+++ pages/api/mutations/use-accept-card/index.ts
@@ -5,7 +5,7 @@ import { UseGetApprovedCardsKey } from '@pages/api/queries/use-get-approved-card
 import { UseGetUnapprovedCardsKey } from '@pages/api/queries/use-get-unapproved-cards'
 import { UseGetRequestedCardsKey } from '@pages/api/queries/use-get-requested-cards'
 import { UseGetAllCardsKey } from '@pages/api/queries/use-get-all-cards'
-import { errorToast, successToast } from '@hooks/use-toast'
+import { errorToast, successToast } from '@utils/toasts'
 
 type UseAcceptCardRequest = {
   cardID: number
