+++ pages/api/mutations/use-open-pack/index.ts
@@ -3,7 +3,7 @@ import axios, { AxiosResponse } from 'axios'
 import { POST } from '@constants/http-methods'
 import { UseGetUserPacksKey } from '@pages/api/queries/use-get-user-packs'
 import { UseGetUserCardsKey } from '@pages/api/queries/use-get-user-cards'
-import { errorToast, successToast } from '@hooks/use-toast'
+import { errorToast, successToast } from '@utils/toasts'
 
 type UseOpenPackRequest = {
   packID: number
