+++ pages/api/mutations/use-update-subscription/index.ts
@@ -2,7 +2,7 @@ import { useMutation, useQueryClient } from 'react-query'
 import axios, { AxiosResponse } from 'axios'
 import { POST } from '@constants/http-methods'
 import { UseGetUserKey } from '@pages/api/queries/use-get-user'
-import { errorToast, successToast } from '@hooks/use-toast'
+import { errorToast, successToast } from '@utils/toasts'
 
 type UseUpdateSubscriptionRequest = {
   uid: number
