+++ pages/open-packs/index.tsx
@@ -1,6 +1,6 @@
 import OpenPackModal from '@components/modals/open-pack-modal'
 import packsMap, { packInfo } from '@constants/packs-map'
-import { warningToast } from '@hooks/use-toast'
+import { warningToast } from '@utils/toasts'
 import useOpenPack from '@pages/api/mutations/use-open-pack'
 import { useGetUser } from '@pages/api/queries'
 import useGetUserPacks from '@pages/api/queries/use-get-user-packs'
