+++ pages/pack-shop/index.tsx
@@ -3,7 +3,7 @@ import getUidFromSession from '@utils/get-uid-from-session'
 import React, { useEffect, useState } from 'react'
 import { packs, packInfo } from '@constants/packs-map'
 import BuyPackModal from '@components/modals/buy-pack-modal'
-import { warningToast } from '@hooks/use-toast'
+import { warningToast } from '@utils/toasts'
 import subscriptionOptions from '@constants/subscription-options'
 import { useGetUser } from '@pages/api/queries'
 import useUpdateSubscription from '@pages/api/mutations/use-update-subscription'
